fix: renamed utils
